+++ st2actions/tests/unit/base.py
@@ -15,12 +15,33 @@
 
 import traceback
 
+import unittest2
+
 from st2tests import config as test_config
 test_config.parse_args()
 
 from st2actions import scheduler, worker, notifier
 from st2common.constants import action as action_constants
 from st2common.models.db.liveaction import LiveActionDB
+from st2common.util.api import get_full_public_api_url
+from st2common.constants.runners import COMMON_ACTION_ENV_VARIABLES
+
+__all__ = [
+    'RunnerTestCase',
+    'MockLiveActionPublisher'
+]
+
+
+class RunnerTestCase(unittest2.TestCase):
+    def assertCommonSt2EnvVarsAvailableInEnv(self, env):
+        """
+        Method which asserts that the common ST2 environment variables are present in the provided
+        environment.
+        """
+        for var_name in COMMON_ACTION_ENV_VARIABLES:
+            self.assertTrue(var_name in env)
+
+        self.assertEqual(env['ST2_ACTION_API_URL'], get_full_public_api_url())
 
 
 class MockLiveActionPublisher(object):