+++ st2actions/tests/unit/policies/test_concurrency.py
@@ -13,7 +13,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import eventlet
 import mock
 import six
 
@@ -53,15 +52,17 @@ LOADER = FixturesLoader()
 FIXTURES = LOADER.load_fixtures(fixtures_pack=PACK, fixtures_dict=TEST_FIXTURES)
 NON_EMPTY_RESULT = 'non-empty'
 
-
-def mock_run(action_parameters):
-    eventlet.sleep(2)
-    return (action_constants.LIVEACTION_STATUS_SUCCEEDED, NON_EMPTY_RESULT, None)
+SCHEDULED_STATES = [
+    action_constants.LIVEACTION_STATUS_SCHEDULED,
+    action_constants.LIVEACTION_STATUS_RUNNING,
+    action_constants.LIVEACTION_STATUS_SUCCEEDED
+]
 
 
 @mock.patch.object(
     TestRunner, 'run',
-    mock.MagicMock(side_effect=mock_run))
+    mock.MagicMock(
+        return_value=(action_constants.LIVEACTION_STATUS_RUNNING, NON_EMPTY_RESULT, None)))
 @mock.patch.object(
     CUDPublisher, 'publish_update',
     mock.MagicMock(side_effect=MockLiveActionPublisher.publish_update))
@@ -96,13 +97,16 @@ class ConcurrencyPolicyTest(EventletTestCase, DbTestCase):
 
     def test_over_threshold(self):
         policy_db = Policy.get_by_ref('wolfpack.action-1.concurrency')
+        self.assertGreater(policy_db.parameters['threshold'], 0)
 
         for i in range(0, policy_db.parameters['threshold']):
             liveaction = LiveActionDB(action='wolfpack.action-1', parameters={'actionstr': 'foo'})
-            eventlet.spawn(action_service.request, liveaction)
+            action_service.request(liveaction)
 
-        # Sleep here to let the threads above schedule the action execution.
-        eventlet.sleep(1)
+        scheduled = LiveAction.get_all()
+        self.assertEqual(len(scheduled), policy_db.parameters['threshold'])
+        for liveaction in scheduled:
+            self.assertIn(liveaction.status, SCHEDULED_STATES)
 
         # Execution is expected to be delayed since concurrency threshold is reached.
         liveaction = LiveActionDB(action='wolfpack.action-1', parameters={'actionstr': 'foo'})
@@ -110,9 +114,10 @@ class ConcurrencyPolicyTest(EventletTestCase, DbTestCase):
         liveaction = LiveAction.get_by_id(str(liveaction.id))
         self.assertEqual(liveaction.status, action_constants.LIVEACTION_STATUS_DELAYED)
 
-        # Sleep here to let the threads above complete the action execution.
-        eventlet.sleep(2)
+        # Mark one of the execution as completed.
+        action_service.update_status(
+            scheduled[0], action_constants.LIVEACTION_STATUS_SUCCEEDED, publish=True)
 
         # Execution is expected to be rescheduled.
         liveaction = LiveAction.get_by_id(str(liveaction.id))
-        self.assertEqual(liveaction.status, action_constants.LIVEACTION_STATUS_RUNNING)
+        self.assertIn(liveaction.status, SCHEDULED_STATES)